Add JWT user extractor and cleanup controllers
Introduce JwtUserExtractor utility to centralize extracting user identifiers from Kong headers and JWT tokens. Update TemplatesController to use JwtUserExtractor and remove the now-duplicated extractUser method. Add explanatory architectural comments to frontend code-tables and notification controllers to document the intentional route duplication for gateway-driven auth. Also remove the deprecated findStuck method from NotificationService.
